requirements.txt
fastapi
uvicorn
python-dotenv
requests
pypdf2
pandas
tqdm
chromadb
langchain
langchain-google-genai
langchain[anthropic]
